Bug 52376 - StatCalculator#addValue(T val, int sampleCount) should use long, not int
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1221465 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: d493569a4fdcf4bd3bd4645f9003a2e98e1ce21a

+++ src/jorphan/org/apache/jorphan/math/StatCalculator.java
@@ -95,7 +95,7 @@ public abstract class StatCalculator<T extends Number & Comparable<? super T>> {
 
     public void addAll(StatCalculator<T> calc) {
         for(Entry<T, MutableLong> ent : calc.valuesMap.entrySet()) {
-            addValue(ent.getKey(), ent.getValue().intValue());
+            addValue(ent.getKey(), ent.getValue().longValue());
         }
     }
 
@@ -197,7 +197,7 @@ public abstract class StatCalculator<T extends Number & Comparable<? super T>> {
 
     protected abstract T divide(T val, int n);
 
-    public void addValue(T val, int sampleCount) {
+    public void addValue(T val, long sampleCount) {
         count += sampleCount;
         double currentVal = val.doubleValue();
         sum += currentVal * sampleCount; 
@@ -215,10 +215,10 @@ public abstract class StatCalculator<T extends Number & Comparable<? super T>> {
     }
 
     public void addValue(T val) {
-        addValue(val,1);
+        addValue(val, 1L);
     }
 
-    private void updateValueCount(T actualValue, int sampleCount) {
+    private void updateValueCount(T actualValue, long sampleCount) {
         MutableLong count = valuesMap.get(actualValue);
         if (count != null) {
             count.add(sampleCount);
